Call parent window from iframe #30
Fixes the test that was failing in Firefox for #30. The old test was
waiting for the iframe to load and then setting a function on the
iframe's window. But Firefox would call `load` after steal had
completed, so it was unreliable. The fix is to set the function on the
parent window and have the iframe call the `parent.synReady()` function.
This guarantees that the function exists before the iframe script has
finished running. Should close out #30