Adds an AMD build
This adds an AMD build that is tested with RequireJS. Instead of running
off of the standalone pluginified tests, this actually pluginifies the
tests to AMD and loads those with RequireJS (which in turn loads Syn
itself). So we have end-to-end RequireJS, ensuring that the build is
compatible. Fixes #61

Depends on #59 and #62